requirements.txt
inflect>=4.1.0
discord>=1.5.1
python-dotenv>=0.14.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib